refactor to break logic up into related areas
