+++ includes/csv_arrow_ingest.hpp
@@ -1,0 +1,9 @@
+#include <arrow/csv/api.h>
+
+std::shared_ptr<arrow::Table>
+ReadEncryptedCsv(const std::string &filename, const std::string *decryption_key,
+                 std::vector<std::string>* utf8_columns);
+
+std::shared_ptr<arrow::Table>
+ReadUnencryptedCsv(const std::string &filename,
+                   std::vector<std::string>* utf8_columns);
